feat(case-control): pass cohort sizes to adjacent component
