+++ src/Analysis/GWASWizard/CovariateStatsByHareCC.jsx
@@ -1,10 +1,10 @@
 /* eslint-disable camelcase */
-import React, { useEffect } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import { useQueries } from 'react-query';
 import Dropdown from '@gen3/ui-component/dist/components/Dropdown';
 import { Spin } from 'antd';
-import { fetchConceptStatsByHare, queryConfig } from './wizardEndpoints/cohortMiddlewareApi';
+import { fetchConceptStatsByHare, queryConfig, getAllHareItems } from './wizardEndpoints/cohortMiddlewareApi';
 
 const CovariateStatsByHareCC = ({
   selectedHare,
@@ -58,12 +58,6 @@ const CovariateStatsByHareCC = ({
     dataControl: results[1].data,
   };
 
-  const getAllHareItems = (concept_value, allCaseHares, allControlHares) => {
-    const caseHareBreakdown = allCaseHares.find((hare) => hare.concept_value === concept_value);
-    const controlHareBreakdown = allControlHares.find((hare) => hare.concept_value === concept_value);
-    return [caseHareBreakdown, controlHareBreakdown];
-  };
-
   const getHareDescriptionBreakdown = (singleHare, allCaseHares, allControlHares) => {
     const hareBreakdown = getAllHareItems(singleHare.concept_value,
       allCaseHares, allControlHares);
@@ -73,23 +67,6 @@ const CovariateStatsByHareCC = ({
     };
     return `${selectedCaseHare.concept_value_name} (sizes: ${selectedCaseHare.persons_in_cohort_with_value}, ${selectedControlHare.persons_in_cohort_with_value})`;
   };
-
-  const handleCohortOverlapProps = (singleHare, allCaseHares, allControlHares) => {
-    const hareBreakdown = getAllHareItems(singleHare.concept_value,
-      allCaseHares, allControlHares);
-    const { caseSize, controlSize } = {
-      caseSize: hareBreakdown[0].persons_in_cohort_with_value,
-      controlSize: hareBreakdown[1].persons_in_cohort_with_value,
-    };
-    return [caseSize, controlSize];
-  };
-
-  useEffect(() => {
-    if (selectedHare && dataCase?.concept_breakdown && dataControl?.concept_breakdown) {
-      //  handleCohortOverlapProps(selectedHare, dataCase.concept_breakdown, dataControl.concept_breakdown)
-      handleHareChange(selectedHare);
-    }
-  }, [selectedHare, dataCase, dataControl, handleHareChange]);
 
   if (statusCase === 'loading' || statusControl === 'loading') {
     return <Spin />;
@@ -126,7 +103,7 @@ const CovariateStatsByHareCC = ({
                 <Dropdown.Item
                   key={`key-${hare.concept_value}`}
                   value={`${hare}`}
-                  onClick={() => handleHareChange(hare)}
+                  onClick={() => handleHareChange(hare, dataCase.concept_breakdown, dataControl.concept_breakdown)}
                 >
                   <div>
                     {getHareDescriptionBreakdown(hare, dataCase.concept_breakdown, dataControl.concept_breakdown)}
